// local-hobbies-frontend/src/types/user.ts
export enum OnboardingSectionName {
    Name = 'name',
    BirthDate = 'birthDate',
    Location = 'location',
    PublicContactInfo = 'publicContactInfo',
    GenderMatched = 'genderMatched',
    ShowAge = 'showAge',
    ShowGenderDisplayed = 'showGenderDisplayed',
    Hobbies = 'hobbies',
    Availabilities = 'availabilities',
}

/**
 * Way in which an onboarding section is incomplete
 */
export enum IncompleteReason {
    NoValue = 'No value',
    MinCountNotMet = 'Minimum number of entries not met',
}

/**
 * Onboarding types
 */

export interface UserOnboardingStatusResponse {
    incompleteSections: UserOnboardingIncompleteSection[];
}

export interface UserOnboardingIncompleteSection {
    name: OnboardingSectionName;
    reason: IncompleteReason;
}